+++ Engine/Runtime/Allocator/source/MallocMimalloc.cpp
@@ -28,12 +28,12 @@ void* MallocMimalloc::Alloc(size_t size)
     return newPtr;
 }
 
-void* MallocMimalloc::AlignedAlloc(size_t size, uint32_t alignment)
+void* MallocMimalloc::AlignedAlloc(size_t size, size_t alignment)
 {
 	void* newPtr = nullptr;
 #if OS_MACOS | OS_IOS
 	// macOS expects all allocations to be aligned to 16 bytes
-	alignment = std::max((uint32_t)16, alignment);
+	alignment = std::max((size_t)16, alignment);
 #else
 #endif
 	newPtr = mi_malloc_aligned(size, alignment);
